refactor(agui): relocate FrontendToolCallRequested + InterruptedWithPayload
Both classes have AG-UI as their sole consumer (FrontendToolCallRequested
via FrontendToolCallRequestedMapper; InterruptedWithPayload via
InterruptedMapper). Symmetric with FrontendStateMutated, which already
lives in langgraph_events.agui._events.

Moves both classes from src/langgraph_events/_event.py to
src/langgraph_events/agui/_events.py and drops the top-level re-exports.
Public import path is now:

  from langgraph_events.agui import FrontendToolCallRequested
  from langgraph_events.agui import InterruptedWithPayload

Pure relocation — no behavior changes. Tests for both classes move from
tests/test_event.py to tests/test_agui.py.
