+++ src/langgraph_events/__init__.py
@@ -13,12 +13,10 @@ from langgraph_events._event import (
     Command,
     DomainEvent,
     Event,  # noqa: F401 (importable for reducer event_type=Event catch-all)
-    FrontendToolCallRequested,
     Halted,
     HandlerRaised,
     IntegrationEvent,
     Interrupted,
-    InterruptedWithPayload,
     Invariant,
     InvariantViolated,
     MaxRoundsExceeded,
@@ -55,13 +53,11 @@ __all__ = [
     "DomainEvent",
     "EventGraph",
     "EventLog",
-    "FrontendToolCallRequested",
     "GraphState",
     "Halted",
     "HandlerRaised",
     "IntegrationEvent",
     "Interrupted",
-    "InterruptedWithPayload",
     "Invariant",
     "InvariantViolated",
     "LLMStreamEnd",
