+++ src/langgraph_events/agui/_mappers.py
@@ -19,14 +19,16 @@ from ag_ui.core import (
 
 from langgraph_events._event import (
     Event,
-    FrontendToolCallRequested,
     Interrupted,
-    InterruptedWithPayload,
     Resumed,
     SystemPromptSet,
 )
 
-from ._events import FrontendStateMutated
+from ._events import (
+    FrontendStateMutated,
+    FrontendToolCallRequested,
+    InterruptedWithPayload,
+)
 from ._protocols import AGUICustomEvent, AGUISerializable
 
 if TYPE_CHECKING:
